requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
astroid==2.15.6
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.7
exceptiongroup==1.1.3
execnet==1.9.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
outcome==1.2.0
packaging==23.1
platformdirs==3.10.0
playwright==1.37.0
pluggy==1.0.0
pycparser==2.21
pyee==9.0.4
pylint==2.17.5
PySocks==1.7.1
pytest==7.3.1
pytest-base-url==2.0.0
pytest-playwright==0.4.2
pytest-xdist==3.3.1
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
tomlkit==0.12.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.3
webdriver-manager==4.0.0
wrapt==1.15.0
wsproto==1.2.0